Review: Move TickUtils into cartesian folder
